# modules/apps/object/object-admin-rest-impl/rest-openapi.yaml
components:
    schemas:
        ObjectDefinition:
            properties:
                actions:
                    additionalProperties:
                        additionalProperties:
                            type: string
                        type: object
                    readOnly: true
                    type: object
                dateCreated:
                    format: date
                    readOnly: true
                    type: string
                dateModified:
                    format: date
                    readOnly: true
                    type: string
                id:
                    format: int64
                    readOnly: true
                    type: integer
                label:
                    additionalProperties:
                        type: string
                    type: object
                name:
                    type: string
                objectFields:
                    items:
                        $ref: "#/components/schemas/ObjectField"
                    type: array
                objectRelationships:
                    items:
                        $ref: "#/components/schemas/ObjectRelationship"
                    type: array
                panelAppOrder:
                    type: string
                panelCategoryKey:
                    type: string
                pluralLabel:
                    additionalProperties:
                        type: string
                    type: object
                scope:
                    type: string
                status:
                    $ref: "#/components/schemas/Status"
                    readOnly: true
                system:
                    readOnly: true
                    type: boolean
            type: object
        ObjectField:
            properties:
                actions:
                    additionalProperties:
                        additionalProperties:
                            type: string
                        type: object
                    readOnly: true
                    type: object
                id:
                    format: int64
                    readOnly: true
                    type: integer
                indexed:
                    type: boolean
                indexedAsKeyword:
                    type: boolean
                indexedLanguageId:
                    type: string
                label:
                    additionalProperties:
                        type: string
                    type: object
                listTypeDefinitionId:
                    format: int64
                    type: integer
                name:
                    type: string
                required:
                    type: boolean
                type:
                    enum: [BigDecimal, Boolean, Date, Double, Integer, Long, String]
                    type: string
        ObjectRelationship:
            properties:
                actions:
                    additionalProperties:
                        additionalProperties:
                            type: string
                        type: object
                    readOnly: true
                    type: object
                id:
                    format: int64
                    readOnly: true
                    type: integer
                label:
                    additionalProperties:
                        type: string
                    type: object
                name:
                    type: string
                objectDefinitionId1:
                    format: int64
                    type: integer
                objectDefinitionId2:
                    format: int64
                    type: integer
                type:
                    enum: [one_to_many, one_to_one, many_to_many, many_to_one]
                    type: string
        Status:
            properties:
                code:
                    format: int32
                    readOnly: true
                    type: integer
                label:
                    type: string
                label_i18n:
                    type: string
            type: object
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.object.admin.rest.client', and version '1.0.0'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Object"
    version: v1.0
openapi: 3.0.1
paths:
    "/object-definitions":
        get:
            operationId: getObjectDefinitionsPage
            parameters:
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: header
                  name: Accept-Language
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/ObjectDefinition"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/ObjectDefinition"
                                type: array
            tags: ["ObjectDefinition"]
        post:
            operationId: postObjectDefinition
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ObjectDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/ObjectDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
            tags: ["ObjectDefinition"]
    "/object-definitions/{objectDefinitionId}":
        delete:
            operationId: deleteObjectDefinition
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
            tags: ["ObjectDefinition"]
        get:
            operationId: getObjectDefinition
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
            tags: ["ObjectDefinition"]
        patch:
            operationId: patchObjectDefinition
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ObjectDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/ObjectDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
            tags: ["ObjectDefinition"]
        put:
            operationId: putObjectDefinition
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ObjectDefinition"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/ObjectDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ObjectDefinition"
            tags: ["ObjectDefinition"]
...
    "/object-definitions/{objectDefinitionId}/publish":
        post:
            operationId: postObjectDefinitionPublish
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json: {}
                        application/xml: {}
            tags: ["ObjectDefinition"]
...